chore: remove router proxy and create error handler
